+++ packages/webauthn/src/index.ts
@@ -1,3 +1,4 @@
+export * from './client-response.js'
 export * from './plugin.js'
 export * from './types.js'
 export { validateAuthentication, validateRegistration } from './validation.js'